requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
# backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==22.8.0
bleach==5.0.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.15
django-extensions==3.2.0
djangorestframework==3.13.1
drf-api-logger==1.1.10
executing==1.1.0
idna==3.3
ipython==8.5.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
numpy==1.23.2
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
phonenumbers==8.12.53
pickleshare==0.7.5
pip-autoremove==0.10.0
pipdeptree==2.2.1
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
tomli==2.0.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
